Palindrome Checker
Checks if a word is palindrome (for example racecar is racecar spelled backwards). The user enters a string then the algorithm checks the string for a reverse of the string having the same matching letters. If they match, then the string is a palindrome. If not, then the string is not a palindrome. A message is displayed in the console showing the result. The program is console based.